Move httptest into httputil package for consistency with other packages
No functional changes
