+++ pkg/aggregator/BUILD.bazel
@@ -41,7 +41,7 @@ go_test(
     deps = [
         "//:go_default_library",
         "//pkg/apis/aggregator/v1:go_default_library",
-        "//pkg/util/httptest:go_default_library",
+        "//pkg/util/httputil/httptest:go_default_library",
         "//pkg/util/logz:go_default_library",
         "//vendor/github.com/SermoDigital/jose/jws:go_default_library",
         "//vendor/github.com/ash2k/stager:go_default_library",
