+++ pkg/creator/luigi/client_test.go
@@ -10,8 +10,8 @@ import (
 	"testing"
 
 	"github.com/atlassian/voyager/pkg/util"
-	. "github.com/atlassian/voyager/pkg/util/httptest"
 	"github.com/atlassian/voyager/pkg/util/httputil"
+	. "github.com/atlassian/voyager/pkg/util/httputil/httptest"
 	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
