+++ pkg/servicecentral/BUILD.bazel
@@ -44,8 +44,8 @@ go_test(
         "//pkg/apis/creator/v1:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/auth:go_default_library",
-        "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
+        "//pkg/util/httputil/httptest:go_default_library",
         "//pkg/util/pkiutil:go_default_library",
         "//pkg/util/pkiutil/pkitest:go_default_library",
         "//pkg/util/testutil:go_default_library",
